+++ src/controller/escritoresLatinoAmericanosController.js
@@ -1,0 +1,9 @@
+const escritores = require('../model/acervoEscritoresLatinoAmericanos.json');
+
+const getAll = (req, res) => {
+    console.log(req.url);
+
+    res.send(escritores);
+};
+
+module.exports = { getAll };
