+++ src/model/acervoEscritoresLatinoAmericanos.json
@@ -1,0 +1,52 @@
+[
+    {
+       "id": 1,
+       "nome": "Gabriel García Márquez",
+       "nacionalidade": "Colombiano"
+    },
+    {
+       "id": 2,
+       "nome": "Gabriela Mistral",
+       "nacionalidade": "Chilena"
+    },
+    {
+        "id": 3,
+        "nome": "Isabel Allende Llona",
+        "nacionalidade": "Peruana"
+    },
+    {
+        "id": 4,
+        "nome": "Jorge Luis Borges",
+        "nacionalidade": "Argentino"
+    },
+    {
+       "id": 5,
+       "nome": "Juan Rulfo",
+       "nacionalidade": "Mexicano" 
+    },
+    {
+        "id": 6,
+        "nome": "Julio Cortázar",
+        "nacionalidade": "Argentino"
+    },
+    {
+        "id": 7,
+        "nome": "Mario Vargas Llosa",
+        "nacionalidade": "Peruano"
+    },
+    {
+        "id": 8,
+        "nome": "Octavio Paz",
+        "nacionalidade": "Mexicano"
+    },
+    {
+        "id": 9,
+        "nome": "Pablo Neruda",
+        "nacionalidade": "Chileno" 
+    },
+    {
+        "id": 10,
+        "nome": "Liliana Colanzi",
+        "nacionalidade": "Boliviana" 
+    }
+]
