--- src/route/cafeteriasRoute.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const controller = require('../controller/cafeteriasController')
-
-router.get("/", controller.getAll);
-router.get("/cafeterias", controller.getAll);
-
-module.exports = router;
